fix for warnings when building with xcode 4

+++ zkSforce/zkSObject.m
@@ -62,7 +62,7 @@ static NSDateFormatter *dateFormatter, *dateTimeFormatter;
 
 - (id) initFromXmlNode:(zkElement *)node {
 	self = [super init];
-	int i, childCount;
+	NSUInteger i, childCount;
 	Id = [[[node childElement:@"Id"] stringValue] copy];
 	type = [[[node childElement:@"type"] stringValue] copy];
 	fields = [[NSMutableDictionary alloc] init];
